pyproject.toml
[project]
name = "qb-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "langchain>=1.2.12",
    "langchain-openai>=1.1.11",
    "langsmith>=0.7.20",
    "openai-whisper>=20250625",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
